+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".AuthActivity">
+        <activity android:name="com.marco97pa.trackmania.auth.AuthActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
