+++ app/src/main/res/layout/activity_auth.xml
@@ -8,7 +8,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".AuthActivity">
+    tools:context="com.marco97pa.trackmania.auth.AuthActivity">
 
     <TextView
         android:id="@+id/textView"
